requirements.txt
python-dotenv
crewai
duckduckgo-search
